feat(timecards): added timeDomainRange functionality
