+++ package.json
@@ -77,8 +77,8 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.10.2",
-    "@carbon/charts": "^0.30.18",
-    "@carbon/charts-react": "^0.30.18",
+    "@carbon/charts": "^0.32.7",
+    "@carbon/charts-react": "^0.32.7",
     "@carbon/colors": "10.8.0",
     "@carbon/icons-react": "10.11.0",
     "@carbon/layout": "10.7.1",
