+++ src/components/BarChartCard/BarChartCard.story.jsx
@@ -111,6 +111,7 @@ storiesOf('Watson IoT/BarChartCard', module)
           size={size}
           onCardAction={action('onCardAction')}
           availableActions={{ expand: true, range: true }}
+          timeRange="thisYear"
         />
       </div>
     );
