+++ src/components/TimeSeriesCard/TimeSeriesCard.jsx
@@ -26,7 +26,12 @@ import {
 } from '../../utils/cardUtilityFunctions';
 import deprecate from '../../internal/deprecate';
 
-import { generateSampleValues, formatGraphTick, findMatchingAlertRange } from './timeSeriesUtils';
+import {
+  generateSampleValues,
+  formatGraphTick,
+  findMatchingAlertRange,
+  generateChartTimeRange,
+} from './timeSeriesUtils';
 
 const { iotPrefix } = settings;
 
@@ -250,6 +255,7 @@ const TimeSeriesCard = ({
   isLoading,
   timeRange,
   showTimeInGMT,
+  endDate,
   ...others
 }) => {
   const {
@@ -437,77 +443,11 @@ const TimeSeriesCard = ({
     [columnNames, i18n.defaultFilterStringPlaceholdText, timeDataSourceId]
   );
 
+  // TODO: remove in next release
   const ChartComponent = chartType === TIME_SERIES_TYPES.BAR ? StackedBarChart : LineChart;
 
-  // If a timeRange is set, the graph should respect it and only show that timeRange even
-  // if some values are outside of the range
-  const timeRangeDomain = [];
-  if (timeRange) {
-    // end time is always now
-    const endTime = moment();
-    // clone the endTime so that it can be mutated
-    let startTime = moment(endTime);
-    switch (timeRange) {
-      case 'last24Hours':
-        startTime.subtract(24, 'hours');
-        timeRangeDomain.push(startTime.valueOf());
-        timeRangeDomain.push(endTime.valueOf());
-        break;
-      case 'last7Days':
-        startTime.subtract(7, 'days');
-        timeRangeDomain.push(startTime.valueOf());
-        timeRangeDomain.push(endTime.valueOf());
-        break;
-
-      case 'lastMonth':
-        startTime.subtract(30, 'days');
-        timeRangeDomain.push(startTime.valueOf());
-        timeRangeDomain.push(endTime.valueOf());
-        break;
-
-      case 'lastQuarter':
-        startTime.subtract(3, 'months');
-        timeRangeDomain.push(startTime.valueOf());
-        timeRangeDomain.push(endTime.valueOf());
-        break;
-
-      case 'lastYear':
-        startTime.subtract(12, 'months');
-        timeRangeDomain.push(startTime.valueOf());
-        timeRangeDomain.push(endTime.valueOf());
-        break;
-
-      case 'thisWeek':
-        startTime = moment().startOf('week');
-        timeRangeDomain.push(startTime.valueOf());
-        timeRangeDomain.push(endTime.valueOf());
-        break;
-
-      case 'thisMonth':
-        startTime = moment().startOf('month');
-        timeRangeDomain.push(startTime.valueOf());
-        timeRangeDomain.push(endTime.valueOf());
-        break;
-
-      case 'thisQuarter':
-        startTime = moment().startOf('quarter');
-        timeRangeDomain.push(startTime.valueOf());
-        timeRangeDomain.push(endTime.valueOf());
-        break;
-
-      case 'thisYear':
-        startTime = moment().startOf('year');
-        timeRangeDomain.push(startTime.valueOf());
-        timeRangeDomain.push(endTime.valueOf());
-        break;
-
-      default:
-        startTime.subtract(24, 'hours');
-        timeRangeDomain.push(startTime.valueOf());
-        timeRangeDomain.push(endTime.valueOf());
-        break;
-    }
-  }
+  const endDateRange = endDate ? moment(endDate) : moment();
+  const timeRangeDomain = generateChartTimeRange(timeRange, endDateRange);
 
   return (
     <Card
